backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
apscheduler==3.10.4
numpy==1.26.4
pyyaml==6.0.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
